requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
httpx
requests
validators